デフォルトのヘッダーをログイン時に対応させる
